py/requirements.txt
atomicwrites==1.3.0
attrs==19.2.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
idna==2.8
importlib-metadata==0.23
isodate==0.6.0
more-itertools==7.2.0
owlrl==5.2.1
packaging==19.2
pkginfo==1.5.0.1
pluggy==0.13.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.2
pytest==5.2.1
git+git://github.com/gordom6/rdflib.git#egg=rdflib
# rdflib==5.0.0.dev0
rdflib-jsonld==0.4.0
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
SPARQLWrapper==1.8.4
tqdm==4.36.1
twine==2.0.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
zipp==0.6.0
